Refactor code into subpackages
