+++ src/main/java/pv168/hotelmasters/superhotel/backend/impl/AccomodationManagerImpl.java
@@ -1,11 +1,16 @@
-package pv168.hotelmasters.superhotel;
+package pv168.hotelmasters.superhotel.backend.impl;
+
+import pv168.hotelmasters.superhotel.backend.interfaces.AccomodationManager;
+import pv168.hotelmasters.superhotel.backend.entities.Accommodation;
+import pv168.hotelmasters.superhotel.backend.entities.Guest;
+import pv168.hotelmasters.superhotel.backend.entities.Room;
 
 import java.util.List;
 
 /**
  * @author 445434
  */
-public class AccomodationManagerImpl implements AccomodationManager{
+public class AccomodationManagerImpl implements AccomodationManager {
     @Override
     public void createAccomodation(Accommodation accommodation) {
