+++ src/main/java/pv168/hotelmasters/superhotel/backend/interfaces/GuestManager.java
@@ -1,4 +1,6 @@
-package pv168.hotelmasters.superhotel;
+package pv168.hotelmasters.superhotel.backend.interfaces;
+
+import pv168.hotelmasters.superhotel.backend.entities.Guest;
 
 import pv168.hotelmasters.superhotel.Exceptions.InvalidEntityException;
