+++ src/test/java/pv168/hotelmasters/superhotel/backend/impl/GuestFactory.java
@@ -1,12 +1,13 @@
-package pv168.hotelmasters.superhotel;
+package pv168.hotelmasters.superhotel.backend.impl;
 
 import java.time.LocalDate;
 import java.time.Month;
+import pv168.hotelmasters.superhotel.backend.entities.Guest;
 
 /**
  * @author Gabriela Godiskova
  */
-public class GuestBuilder {
+public class GuestFactory {
 
     private Long id;
     private String name;
@@ -14,27 +15,27 @@ public class GuestBuilder {
     private LocalDate birthday;
     private Long crCardNm;
 
-    public GuestBuilder id (Long id) {
+    public GuestFactory id (Long id) {
         this.id = id;
         return this;
     }
 
-    public GuestBuilder name(String name) {
+    public GuestFactory name(String name) {
         this.name = name;
         return this;
     }
 
-    public GuestBuilder address(String address) {
+    public GuestFactory address(String address) {
         this.address = address;
         return this;
     }
 
-    public GuestBuilder birthday(LocalDate birthday) {
+    public GuestFactory birthday(LocalDate birthday) {
         this.birthday = birthday;
         return this;
     }
 
-    public GuestBuilder birthday(int year, Month month, int day) {
+    public GuestFactory birthday(int year, Month month, int day) {
         this.birthday = LocalDate.of(year,month,day);
         return this;
     }
@@ -49,7 +50,7 @@ public class GuestBuilder {
         return guest;
     }
 
-    public GuestBuilder crCardNm(Long crCardNm) {
+    public GuestFactory crCardNm(Long crCardNm) {
         this.crCardNm = crCardNm;
         return this;
     }
